cleanup: form state no longer needs combined state - separating config and hydrate state - preps refactor to move fields from the rest of the ConfigurationStateProvider - moves HydratedRevisionContext to state folder to access Config utils
