+++ src/components/Configure/UpdateInstallation.tsx
@@ -7,13 +7,13 @@ import {
   ErrorBoundary,
   useErrorState,
 } from '../../context/ErrorContextProvider';
-import { useHydratedRevision } from '../../context/HydratedRevisionContext';
 import { useInstallIntegrationProps } from '../../context/InstallIntegrationContext';
 import { useProject } from '../../context/ProjectContext';
 import { Installation, Integration } from '../../services/api';
 
 import { onSaveUpdate } from './actions/onSaveUpdate';
 import { useConfigureState } from './state/ConfigurationStateProvider';
+import { useHydratedRevision } from './state/HydratedRevisionContext';
 import { getConfigureState, resetConfigurationState } from './state/utils';
 import { ConfigureInstallationBase } from './ConfigureInstallationBase';
 import { useSelectedObjectName } from './ObjectManagementNav';
